# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ownly API
  version: 0.1.0
  description: API starter for the Ownly Personal Ownership Management Platform.
servers:
  - url: https://api.ownly.app/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    UploadSessionCreate:
      type: object
      required: [workspaceId, filename, mimeType, byteSize, documentKind]
      properties:
        workspaceId:
          type: string
          format: uuid
        filename:
          type: string
        mimeType:
          type: string
        byteSize:
          type: integer
        checksumSha256:
          type: string
        documentKind:
          type: string
          enum: [invoice, warranty, manual, receipt, photo, service_record, insurance, other]
    UploadSession:
      type: object
      required: [uploadId, documentId, uploadUrl, expiresAt]
      properties:
        uploadId:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          type: string
          format: date-time
...
paths:
  /auth/register:
    post:
      security: []
      summary: Create a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                fullName:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      security: []
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login result with tokens or MFA challenge
...
  /uploads/sessions:
    post:
      summary: Create a signed encrypted upload session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadSessionCreate'
      responses:
        '201':
          description: Upload session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
  /uploads/{uploadId}/complete:
    post:
      summary: Confirm upload and enqueue scanner pipeline
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Processing started
...